Branches epic git log format into two  versions.
One version is friendly (i.e. it uses friendly formated dates) and uses
a modified order of elements in the log message.  The other format is
exact and uses the ISO 8601 timestamp with a modified order of log
elements.

Lastly, this commit modifes the log related aliases to use these two new
formats.